Return in controller change from type to ResponseEntity. Added Feign and Hystrix.Testing needed

+++ MovieService/src/main/java/rs/uns/acs/ftn/controllers/MovieController.java
@@ -1,9 +1,12 @@
 package rs.uns.acs.ftn.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cloud.netflix.feign.FeignClient;
 import org.springframework.core.env.Environment;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 
 import javax.ws.rs.core.MediaType;
 import org.springframework.transaction.annotation.Transactional;
@@ -43,15 +46,20 @@ public class MovieController extends AbstractRESTController<Movie, String> {
 		this.movieService = movieService;
 	}
 	
-	/* TODO: Administrator verification. */
+	/* TODO: TEST ADMIN */
 	@Transactional
 	@RequestMapping(value = "/addMovie",
 			method = RequestMethod.POST,
 			consumes = MediaType.APPLICATION_JSON,
 			produces = MediaType.APPLICATION_JSON)
-	public Movie addMovie(@RequestBody MovieDTO dto){
-			Movie movie = movieService.addMovie(dto.getName(), dto.getDescription(), dto.getLength(), dto.getDirector(), dto.getActors(), dto.getCategory(), dto.getRatings(), dto.getPremiere());
-			return movie;
+	public ResponseEntity<Movie> addMovie(@RequestBody MovieDTO dto, @RequestParam("sessionId") String sessionId){
+			Movie movie = new Movie();
+			if (!isAdmin(movieService.getTypeBySessionIdLoad(sessionId))){
+				return new ResponseEntity<Movie>(movie, HttpStatus.FORBIDDEN);			
+			}
+		
+			movie = movieService.addMovie(dto.getName(), dto.getDescription(), dto.getLength(), dto.getDirector(), dto.getActors(), dto.getCategory(), dto.getRatings(), dto.getPremiere());
+			return new ResponseEntity<Movie>(movie, HttpStatus.OK);
 	}
 	
 	
@@ -71,54 +79,66 @@ public class MovieController extends AbstractRESTController<Movie, String> {
 	@RequestMapping(value = "/getMovie",
 			method = RequestMethod.GET,
 			produces = MediaType.APPLICATION_JSON)
-	public Movie getMovie(@RequestParam(name = "movie_id") String movieId){
+	public ResponseEntity<Movie> getMovie(@RequestParam(name = "movie_id") String movieId){
 			Movie movie = movieService.findById(movieId);
-			return movie;
+			return new ResponseEntity<Movie>(movie, HttpStatus.OK);
 	}
 	
-	/* TODO: Administrator verification. */
+	/* TODO: TEST ADMIN */
 	@Transactional
 	@RequestMapping(value = "/updateMovie",
 			method = RequestMethod.POST,
 			consumes = MediaType.APPLICATION_JSON,
 			produces = MediaType.APPLICATION_JSON)
-	public Movie updateMovie(@RequestBody MovieDTO dto){
-			Movie movie = movieService.updateMovie(dto.getId(), dto.getName(), dto.getDescription(), dto.getLength(), dto.getDirector(), dto.getActors(), dto.getCategory(), dto.getRatings(), dto.getPremiere());
-			return movie;
+	public ResponseEntity<Movie> updateMovie(@RequestBody MovieDTO dto, @RequestParam("sessionId") String sessionId){
+			Movie movie = new Movie();
+			if (!isAdmin(movieService.getTypeBySessionIdLoad(sessionId))){
+				return new ResponseEntity<Movie>(movie, HttpStatus.FORBIDDEN);			
+			}
+			movie = movieService.updateMovie(dto.getId(), dto.getName(), dto.getDescription(), dto.getLength(), dto.getDirector(), dto.getActors(), dto.getCategory(), dto.getRatings(), dto.getPremiere());
+			return new ResponseEntity<Movie>(movie, HttpStatus.OK);
 	}
 	
-	/* TODO: Administrator verification. */
+	/* TODO: TEST ADMIN */
 	@Transactional
 	@RequestMapping(value = "/deleteMovie",
 			method = RequestMethod.POST,
 			consumes = MediaType.TEXT_PLAIN,
 			produces = MediaType.APPLICATION_JSON)
-	public Boolean deleteMovie(@RequestBody String movieId){
+	public ResponseEntity<Boolean> deleteMovie(@RequestBody String movieId, @RequestParam("sessionId") String sessionId){
+
+			if (!isAdmin(movieService.getTypeBySessionIdLoad(sessionId))){
+				return new ResponseEntity<Boolean>(false, HttpStatus.FORBIDDEN);			
+			}
 			Boolean removed = movieService.delete(movieId);
-			return removed;
+			return new ResponseEntity<Boolean>(removed, HttpStatus.OK);
 	}
 	
-	/* TODO: User verification. */
+	/* TODO: TEST USER */
 	@Transactional
 	@RequestMapping(value = "/rateMovie",
 			method = RequestMethod.POST,
 			consumes = MediaType.APPLICATION_JSON,
 			produces = MediaType.APPLICATION_JSON)
-	public Movie rateMovie(@RequestBody RatingDTO dto){
+	public ResponseEntity<Movie> rateMovie(@RequestBody RatingDTO dto, @RequestParam("sessionId") String sessionId){
+			Movie movie = new Movie();
+			if (!isAdmin(movieService.getTypeBySessionIdLoad(sessionId))){
+				return new ResponseEntity<Movie>(movie, HttpStatus.FORBIDDEN);			
+			}
 			Rating rating = new Rating(dto.getValue(), dto.getPosterId());
-			Movie movie = movieService.rateMovie(dto.getMovieId(), rating);
-			return movie;
+			movie = movieService.rateMovie(dto.getMovieId(), rating);
+			return new ResponseEntity<Movie>(movie, HttpStatus.OK);
 	}
 	
 	@Transactional
 	@RequestMapping(value = "/getMoviesByYear",
 			method = RequestMethod.GET,
 			produces = MediaType.APPLICATION_JSON)
-	public List<Movie> getMoviesByYear(@RequestParam(name = "year_str") String  yearStr){
+	public ResponseEntity<List<Movie>> getMoviesByYear(@RequestParam(name = "year_str") String  yearStr){
 			try {
 				Integer year = new Integer(yearStr);
 				List<Movie> movies = movieService.findByYear(year);
-				return movies;
+				return new ResponseEntity<List<Movie>>(movies, HttpStatus.OK);
 			} catch (NumberFormatException e){
 				e.printStackTrace();
 				return null;
@@ -129,32 +149,32 @@ public class MovieController extends AbstractRESTController<Movie, String> {
 	@RequestMapping(value = "/getTrending",
 			method = RequestMethod.GET,
 			produces = MediaType.APPLICATION_JSON)
-	public List<Movie> getTrending(){
+	public ResponseEntity<List<Movie>> getTrending(){
 		Integer year = Calendar.getInstance().get(Calendar.YEAR);
 
 			List<Movie> movies = movieService.findByYear(year);
-			return movies;
+			return new ResponseEntity<List<Movie>>(movies, HttpStatus.OK);
 	}
 	
 	@Transactional
 	@RequestMapping(value = "/getByRating",
 			method = RequestMethod.GET,
 			produces = MediaType.APPLICATION_JSON)
-	public List<Movie> getByRating(@RequestParam(name = "rating") String rating){
+	public ResponseEntity<List<Movie>> getByRating(@RequestParam(name = "rating") String rating){
 			double ratingAvg = new Double(rating);
 			List<Movie> movies = movieService.findByRatingAvg(ratingAvg);
-			return movies;
+			return new ResponseEntity<List<Movie>>(movies, HttpStatus.OK);
 	}
 	
 	@Transactional
 	@RequestMapping(value = "/getByDirector",
 			method = RequestMethod.GET,
 			produces = MediaType.APPLICATION_JSON)
-	public List<Movie> getByDirector(@RequestParam(name = "director_name") String directorName,
+	public ResponseEntity<List<Movie>> getByDirector(@RequestParam(name = "director_name") String directorName,
 			  					     @RequestParam(name = "director_surname") String directorSurname){
 
 			List<Movie> movies = movieService.findByDirector(directorName, directorSurname);
-			return movies;
+			return new ResponseEntity<List<Movie>>(movies, HttpStatus.OK);
 
 	}
 	
@@ -162,10 +182,10 @@ public class MovieController extends AbstractRESTController<Movie, String> {
 	@RequestMapping(value = "/getByActor",
 			method = RequestMethod.GET,
 			produces = MediaType.APPLICATION_JSON)
-	public List<Movie> getByActor(@RequestParam(name = "actor_name") String actorName,
+	public ResponseEntity<List<Movie>> getByActor(@RequestParam(name = "actor_name") String actorName,
 			 					  @RequestParam(name = "actor_surname") String actorSurname){
 			List<Movie> movies = movieService.findByActor(actorName, actorSurname);
-			return movies;
+			return new ResponseEntity<List<Movie>>(movies, HttpStatus.OK);
 
 	}
 	
@@ -173,11 +193,11 @@ public class MovieController extends AbstractRESTController<Movie, String> {
 	@RequestMapping(value = "/getByCategory",
 			method = RequestMethod.GET,
 			produces = MediaType.APPLICATION_JSON)
-	public List<Movie> getByCategory(@RequestParam(name = "category") String category){
+	public ResponseEntity<List<Movie>> getByCategory(@RequestParam(name = "category") String category){
 
 			Movie.Category cat = Movie.Category.valueOf(category);
 			List<Movie> movies = movieService.findByCategory(cat);
-			return movies;
+			return new ResponseEntity<List<Movie>>(movies, HttpStatus.OK);
 	}
 	
 	@Transactional
@@ -185,10 +205,10 @@ public class MovieController extends AbstractRESTController<Movie, String> {
 			method = RequestMethod.POST,
 			consumes = MediaType.APPLICATION_JSON,
 			produces = MediaType.APPLICATION_JSON)
-	public List<Movie> getByActors(@RequestBody List<Actor> actors){
+	public ResponseEntity<List<Movie>> getByActors(@RequestBody List<Actor> actors){
 
 			List<Movie> movies = movieService.findByActors(actors);
-			return movies;
+			return new ResponseEntity<List<Movie>>(movies, HttpStatus.OK);
 
 	}
 	
@@ -197,10 +217,38 @@ public class MovieController extends AbstractRESTController<Movie, String> {
 			method = RequestMethod.POST,
 			consumes = MediaType.TEXT_PLAIN,
 			produces = MediaType.TEXT_PLAIN)
-	public String getMovieName(@RequestBody String movieId){
-
-			return movieService.findMovieName(movieId);
+	public ResponseEntity<String> getMovieName(@RequestBody String movieId){
+			String name = movieService.findMovieName(movieId);
+			return new ResponseEntity<String>(name, HttpStatus.OK) ;
 
 	}
+	
+	@FeignClient("user-service") // the application.name for the user service
+	public interface UserServiceClient {
+		@RequestMapping(value = "users/get_type_by_session_id", method = RequestMethod.GET) // the endpoint which will be balanced over
+		
+		String getTypeBySessionId(@RequestParam(name = "sessionId") String sessionId);// the method specification must
+																						// be the same as for
+																						// users/get_type_by_session_id
+	}
+	
+	
+	/**
+	 * Method compare if given string is equal to "ADMINISTRATOR"
+	 * 
+	 * @param type
+	 * @return if user type is administrator
+	 */
+	private boolean isAdmin(String type) {
+		if (type == null) {
+			return false;
+		}
+
+		if (type.equals("ADMINISTRATOR")) {
+			return true;
+		}
+
+		return false;
+	}	
 
 }
